ai_service_fastapi/requirements.txt
# FastAPI AI Service Dependencies

# FastAPI Core
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# Pydantic
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
databases[postgresql]>=0.9.0
asyncpg>=0.29.0

# ChromaDB & Embeddings
chromadb>=0.4.22
langchain-chroma>=0.1.0
langchain-community>=0.0.20
langchain>=0.1.0
sentence-transformers>=2.2.2

# AI
google-generativeai>=0.3.2

# RabbitMQ
aio-pika>=9.3.0

# Caching
cachetools>=5.3.2

# Utilities
python-dotenv>=1.0.0
httpx>=0.26.0
